+++ src/RcsbSeacrh/FacetStore/FacetMemberInterface.ts
@@ -1,10 +1,16 @@
 
 import {ChartConfigInterface, ChartType} from "@rcsb/rcsb-charts/lib/RcsbChartComponent/ChartConfigInterface";
 import {
-    AttributeFacetType,
+    AttributeFacetType, BucketDataType,
     FilterFacetType
 } from "@rcsb/rcsb-search-tools/lib/SearchParseTools/SearchFacetInterface";
 import {RcsbSearchAttributeType} from "@rcsb/rcsb-api-tools/build/RcsbSearch/Types/SearchMetadata";
+import {SearchQueryType} from "@rcsb/rcsb-search-tools/lib/SearchQueryTools/SearchQueryInterfaces";
+import {
+    ChartDataColumnInterface,
+    ChartDataValueInterface
+} from "@rcsb/rcsb-charts/lib/RcsbChartDataProvider/ChartDataProviderInterface";
+import React from "react";
 
 export type FacetType = AttributeFacetType  | FilterFacetType;
 
@@ -14,7 +20,14 @@ export interface FacetMemberInterface {
     attributeName: string;
     attribute: RcsbSearchAttributeType;
     chartType: ChartType;
-    chartConfig?: ChartConfigInterface & { mergeDomainMaxValue?:number; };
+    chartConfig?: ChartConfigInterface;
+    facetConfig?: FacetConfigInterface;
     facet: FacetType;
     contentType:"date"|"number"|"string";
+}
+
+export interface FacetConfigInterface {
+    mergeDomainMaxValue?: number;
+    facetTransform?: (d: BucketDataType[])=>BucketDataType[];
+    bucketClickSearchQuery?: (datum:ChartDataValueInterface, data: ChartDataColumnInterface[], e?: React.MouseEvent)=>SearchQueryType;
 }